+++ src/features/dashboard/components/index.ts
@@ -1,11 +1,11 @@
 export { DashboardKpiCards } from "./DashboardKpiCards";
 export { DashboardPettyCashRequests } from "./DashboardPettyCashRequests";
-export { DashboardSalesChart } from "./DashboardSalesChart";
 export { DashboardRecentTransactions } from "./DashboardRecentTransactions";
 export { DashboardTransactionLog } from "./DashboardTransactionLog";
 export { DashboardExpenseChart } from "./DashboardExpenseChart";
 export { DashboardCashflowChart } from "./DashboardCashflowChart";
-export { Dashboard30DayChart } from "./Dashboard30DayChart";
+export { DashboardComparisonChart } from "./DashboardComparisonChart";
+export { DashboardKpiTrendChart } from "./DashboardKpiTrendChart";
 export { DashboardKpiRichGrid } from "./DashboardKpiRichGrid";
 export { DashboardAnalysisDrill } from "./DashboardAnalysisDrill";
 export { DashboardTopProducts } from "./DashboardTopProducts";